backend/requirements.txt
# FastAPI Core
fastapi>=0.115.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy>=2.0.35
psycopg2-binary==2.9.9
alembic==1.12.1

# Validation
pydantic>=2.5.0,<3.0.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Reports
reportlab==4.0.7
openpyxl==3.1.2
pandas==2.1.3

# Images
Pillow==10.1.0

# CORS & Environment
python-dotenv==1.0.0

# Date handling
python-dateutil==2.8.2
# Email
fastapi-mail 
# Payment
paypalrestsdk